refactoring(file): 파일 디렉토리 정리
