+++ src/pages/article/core/abstractView.ts
